Fix some warnings
