+++ src/common/scripting/backend/codegen.cpp
@@ -21,19 +21,16 @@
 **
 */
 
-#include <stdlib.h>
+#include "basics.h"
 #include "cmdlib.h"
 #include "codegen.h"
-#include "sc_man.h"
-#include "v_text.h"
 #include "filesystem.h"
-#include "v_video.h"
-#include "utf8.h"
-#include "texturemanager.h"
 #include "m_random.h"
-#include "v_font.h"
 #include "palettecontainer.h"
-
+#include "printf.h"
+#include "sc_man.h"
+#include "texturemanager.h"
+#include "v_font.h"
 
 extern FRandom pr_exrandom;
 FMemArena FxAlloc(65536);
