+++ src/console/c_notifybuffer.cpp
@@ -23,13 +23,14 @@
 */
 
 #include "c_console.h"
-#include "vm.h"
-#include "gamestate.h"
 #include "c_cvars.h"
+#include "c_notifybufferbase.h"
+#include "gamestate.h"
+#include "i_time.h"
+#include "printf.h"
 #include "sbar.h"
 #include "v_video.h"
-#include "i_time.h"
-#include "c_notifybufferbase.h"
+#include "vm.h"
 
 struct FNotifyBuffer : public FNotifyBufferBase
 {
